+++ src/LamaPon/LamaPon.h
@@ -78,6 +78,7 @@
 #include "LamaPon/Graphics/LitMaterial.h"
 #include "LamaPon/Graphics/LitMaterialAsset.h"
 #include "LamaPon/Graphics/ShadowMap.h"
+#include "LamaPon/Graphics/SpriteRendering.h"
 #include "LamaPon/Graphics/TextLayout.h"
 #include "LamaPon/Input/InputSystem.h"
 #include "LamaPon/Physics/Raycast.h"